Update #format_description to use Markdown, rather than #simple_format.
The issue is that Rails' #simple_format helper keeps changing between
versions in problematic ways, and the latest one adds unwanted BR tags
everywhere. This is becoming a maintenance problem.

The solution is to use the well-understood Markdown format for doing
formatting, which does many of the nice things we liked about
#simple_format, plus others. However, users can still enter most HTML tags
if they don't want to learn Markdown.